src/requirements.txt
python-dotenv==1.0.0
functions-framework==3.*
google-cloud-aiplatform